Avoid overwritting files when renaming (don't rename if a file with the same name already exists)

+++ pyrenamer/src/pyrenamerfilefuncs.py
@@ -430,6 +430,10 @@ def replace_music(name, path, newname, newpath):
 
 def rename_file(ori, new):
     """ Change filename with the new one """
+    if os.path.exists(new):
+        print "Error while renaming %s to %s! -> %s already exists!" % (ori, new, new)
+        return False
+    
     try:
         os.rename(ori, new)
         print "Renaming %s to %s" % (ori, new)
